Add yarn run v1.21.1 $ react-scripts build && scp -r build/* novickjd.dev:/var/www/piday Creating an optimized production build... Compiled successfully.
File sizes after gzip:

  53.74 KB  build/static/js/2.06dd2869.chunk.js
  6.63 KB   build/static/js/main.b1cca69a.chunk.js
  774 B     build/static/js/runtime-main.334298aa.js
  470 B     build/static/css/main.d8857a9a.chunk.css

The project was built assuming it is hosted at /.
You can control this with the homepage field in your package.json.

The build folder is ready to be deployed.
You may serve it with a static server:

  yarn global add serve
  serve -s build

Find out more about deployment here:

  bit.ly/CRA-deploy

Done in 11.68s. command